Split Twitter dependency into it's separate module
