+++ src/main/java/tweeting/injection/modules/Twitter4JModule.java
@@ -2,15 +2,9 @@ package tweeting.injection.modules;
 
 import dagger.Module;
 import dagger.Provides;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import tweeting.conf.TweetingConfiguration;
-import tweeting.conf.TwitterOAuthCredentials;
 import tweeting.services.Twitter4JService;
 import tweeting.services.TwitterService;
 import twitter4j.Twitter;
-import twitter4j.TwitterFactory;
-import twitter4j.conf.ConfigurationBuilder;
 
 import javax.inject.Singleton;
 
@@ -20,31 +14,8 @@ import javax.inject.Singleton;
  * Note: TweetingConfiguration is required for configuring the API the Twitter4JService relies on.
  */
 
-@Module
+@Module(includes = TwitterModule.class)
 public class Twitter4JModule {
-
-    private static final Logger logger = LoggerFactory.getLogger(Twitter4JModule.class);
-
-    @Provides
-    @Singleton
-    Twitter provideTwitterAPI(TweetingConfiguration conf) {
-        if (conf.getTwitterAuthorization() == null) {
-            logger.error("TwitterService was built without valid " +
-                    "Twitter Credentials. Twitter Credentials were null.", new NullPointerException());
-            return TwitterFactory.getSingleton();
-        }
-        final TwitterOAuthCredentials auth = conf.getTwitterAuthorization();
-        return new TwitterFactory(
-                new ConfigurationBuilder()
-                        .setDebugEnabled(true)
-                        .setJSONStoreEnabled(true) // Need in order to use getRawJSON
-                        .setOAuthConsumerKey(auth.getConsumerAPIKey())
-                        .setOAuthConsumerSecret(auth.getConsumerAPISecretKey())
-                        .setOAuthAccessToken(auth.getAccessToken())
-                        .setOAuthAccessTokenSecret(auth.getAccessTokenSecret())
-                        .build())
-                .getInstance();
-    }
 
     @Provides
     @Singleton
